Re-org presets, ops, UI
