+++ scripts/addons/cam/tests/gcode_generator.py
@@ -1,20 +1,9 @@
 import sys
 import warnings
-import os
 
 import bpy
 
 warnings.simplefilter("once")
-
-# try:
-#     bpy.context.preferences.addons["bl_ext.user_default.fabex"]
-# except KeyError:
-#     print("Addon is not installed, attempting install...")
-#     bpy.context.preferences.system.use_online_access = True
-#     bpy.ops.extensions.repo_sync_all(use_active_only=False)
-#     bpy.ops.extensions.package_install_files(filepath=f"{sys.argv[1]}", repo="user_default")
-
-sys.path.insert(0, os.path.abspath(".."))
 
 # Get the scene
 s = bpy.context.scene
